test: PATCH /user API 테스트 완료
